task diag: Account for DIAG_VMA size when estimating message size
Currently DIAG_VMA and VMA_NAME are not considered when estimating
message size. This leads to a number of reallocations when filling
in the packet. Ballpark size by counting number of VMAs and
ballparking the average filename with 128 bytes (note some maps
have no filename, others will be longer - this is just a start).

Signed-off-by: David Ahern <dsahern@gmail.com>