task diag: Change implementation of DIAG_VMA attribute
Currently each VMA for a task generates a TASK_DIAG_VMA and
TASK_DIAG_VMA_NAME attribute in the netlink message. This means
the message has multiple of these attributes which breaks the
ability to use libnl for message parsing.

This patch puts the filename into the task_diag_vma struct and
converts TASK_DIAG_VMA attribute into a series of task_diag_vma.
Now is there is a single TASK_DIAG_VMA attribute that is parsed
as:

 | struct task_diag_vma | filename | ...

Signed-off-by: David Ahern <dsahern@gmail.com>