task_diag: Change implementation of DIAG_VMA attribute
Add an ability to add a new fields into the task_diag_vma structures
without breaking backward compatibility.

In addition it allows us to add an optional data about vma-s. For
example, we are going to add memory consumption data, only if they are
requested, because this is a heavy operation.

Signed-off-by: Andrey Vagin <avagin@openvz.org>