[RFC] overlayfs: don't show inode numbers from underlying fs
The first thing is that in this case we can return two equal inode
numbers for different files, if they are saved on differen underlying
file systems.

The second thing is that inode numbers are shown in inotify fdinfo files
and currently stat() and fdinfo reports different values. fdinfo
gets i_ino from overlayfs inodes and stat() gets i_ino from underlying
inodes.

We meet this problem when we try to checkpoint a container on overlayfs.
CRIU read an inode number from an inotify fdinfo file, then walked a
whole file system and failed to find a file with this inode number.

Cc: Miklos Szeredi <miklos@szeredi.hu>