netlink: allow to set peeking offset for sockets
This allows us to read socket's queue without removing skbs from it.

The same logic was implemented for unix sockets and we use this to
dump and restore sockets in CRIU.

I am not sure whether sk_peek_off has to be protected by locks. If it
isn't protected, we don't break existing behaviour, but logic around
sk_peek_off works properly only if a socket is used by one user.

If we add a lock, we will reduce a perfomance.