namei: add LOOKUP_DFD_ROOT to use dfd as root
The problem is that a pathname can contain absolute symlinks and now
they are resolved relative to the current root.

If we want to open a file in another mount namespaces and we have a file
descriptor to its root directory, we probably want to resolve pathname
in the target mount namespace. For this we add this new flag.

If LOOKUP_DFD_ROOT is set, path_init() initializes nd->root and nd->path
to the same value.

Changes since v1:
* initialize nd->root_seq (thanks to Omar Sandoval for reporting and
  fixing this issue)

Cc: Omar Sandoval <osandov@osandov.com>