mnt: allow to add a mount into a shared group
Now a shared group can be only inherited from a source mount.
This patch adds an ability to add a mount into an existing shared
group.

mount(source, target, NULL, MS_SET_GROUP, NULL)

mount() with the MS_SET_GROUP flag adds the "target" mount into a
group of the "source" mount. The process has to have the CAP_SYS_ADMIN
capability in namespaces of these mounts.

This new functionality together with "mnt: Tuck mounts under others
instead of creating shadow/side mounts." allows us to dump and restore
any set of mount namespaces. On the first step we restore a tree of
mounts and on the second step we restore shared and master id-s.

Cc: Eric W. Biederman <ebiederm@xmission.com>
Cc: Alexander Viro <viro@zeniv.linux.org.uk>