unix: report mnt_id for unix sockets via socket_diag
Currently here is no way to understand where a socket is placed.
It is possible to get device and inode numbers, but it isn't possible
to get a mount point.

This patch adds mnt_id into the unix_diag_vfs structure.

In CRIU we need this to dump and restore unix sockets.