unix: report file system path for unix sockets via socket_diag
Currently only an address is reported for a socket, but it can contain a
relative path or a socket file may be moved to somewhere, so now here is
no way to get a real path to a socket file.

This patch expands the socket_diag interface to get a real path to a
socket file. It is reported only if the UDIAG_SHOW_PATH is set.

In CRIU we need these paths to dump and restore unix sockets.